+++ internal/provider/incident_workflow_resource.go
@@ -1,0 +1,215 @@
+package provider
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/incident-io/terraform-provider-incident/internal/apischema"
+	"github.com/incident-io/terraform-provider-incident/internal/client"
+)
+
+var (
+	_ resource.Resource                = &IncidentWorkflowResource{}
+	_ resource.ResourceWithImportState = &IncidentWorkflowResource{}
+)
+
+type IncidentWorkflowResource struct {
+	client *client.ClientWithResponses
+}
+
+func NewIncidentWorkflowResource() resource.Resource {
+	return &IncidentWorkflowResource{}
+}
+
+type IncidentWorkflowResourceModel struct {
+	ID               types.String `tfsdk:"id"`
+	Name             types.String `tfsdk:"name"`
+	Folder           types.String `tfsdk:"folder"`
+	Trigger          types.String `tfsdk:"trigger"`
+	TerraformRepoURL types.String `tfsdk:"terraform_repo_url"`
+}
+
+func (r *IncidentWorkflowResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_workflow"
+}
+
+func (r *IncidentWorkflowResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		MarkdownDescription: apischema.TagDocstring("Workflows V2"),
+		Attributes: map[string]schema.Attribute{
+			"id": schema.StringAttribute{
+				Computed: true,
+			},
+			"name": schema.StringAttribute{
+				Required: true,
+			},
+			"folder": schema.StringAttribute{
+				Optional: true,
+			},
+			"trigger": schema.StringAttribute{
+				Required: true,
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+			},
+			"terraform_repo_url": schema.StringAttribute{
+				Required: true,
+			},
+		},
+	}
+}
+
+func (r *IncidentWorkflowResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var data *IncidentWorkflowResourceModel
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	payload := client.CreateWorkflowRequestBody{
+		Trigger: data.Trigger.ValueString(),
+		Workflow: client.WorkflowPayload{
+			Name:             data.Name.ValueString(),
+			TerraformRepoUrl: data.TerraformRepoURL.ValueStringPointer(),
+			OnceFor:          []string{"incident.url"},
+			ConditionGroups:  []client.ExpressionFilterOptsPayloadV2{},
+			Steps:            []client.StepConfigPayload{},
+			Expressions:      []client.ExpressionPayloadV2{},
+			RunsOnIncidents:  "newly_created",
+			IsDraft:          true,
+		},
+	}
+	if folder := data.Folder.ValueString(); folder != "" {
+		payload.Folder = &folder
+	}
+
+	result, err := r.client.WorkflowsV2CreateWorkflowWithResponse(ctx, payload)
+	if err == nil && result.StatusCode() >= 400 {
+		err = fmt.Errorf(string(result.Body))
+	}
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to create workflow, got error: %s", err))
+		return
+	}
+
+	tflog.Trace(ctx, fmt.Sprintf("created a workflow resource with id=%s", result.JSON201.Workflow.Id))
+	data = r.buildModel(result.JSON201.Workflow)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+func (r *IncidentWorkflowResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var state *IncidentWorkflowResourceModel
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	var data *IncidentWorkflowResourceModel
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	payload := client.WorkflowsV2UpdateWorkflowJSONRequestBody{
+		Workflow: client.WorkflowPayload{
+			Name:             data.Name.ValueString(),
+			TerraformRepoUrl: data.TerraformRepoURL.ValueStringPointer(),
+			OnceFor:          []string{"incident.url"},
+			ConditionGroups:  []client.ExpressionFilterOptsPayloadV2{},
+			Steps:            []client.StepConfigPayload{},
+			Expressions:      []client.ExpressionPayloadV2{},
+			RunsOnIncidents:  "newly_created",
+			IsDraft:          true,
+		},
+	}
+
+	result, err := r.client.WorkflowsV2UpdateWorkflowWithResponse(ctx, state.ID.ValueString(), payload)
+	if err == nil && result.StatusCode() >= 400 {
+		err = fmt.Errorf(string(result.Body))
+	}
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to update workflow, got error: %s", err))
+		return
+	}
+
+	data = r.buildModel(result.JSON200.Workflow)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+func (r *IncidentWorkflowResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var data *IncidentWorkflowResourceModel
+	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	result, err := r.client.WorkflowsV2ShowWorkflowWithResponse(ctx, data.ID.ValueString())
+	if err == nil && result.StatusCode() >= 400 {
+		err = fmt.Errorf(string(result.Body))
+	}
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to read workflow, got error: %s", err))
+		return
+	}
+
+	data = r.buildModel(result.JSON200.Workflow)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+func (r *IncidentWorkflowResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var data *IncidentWorkflowResourceModel
+	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	_, err := r.client.WorkflowsV2DestroyWorkflowWithResponse(ctx, data.ID.ValueString())
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to delete workflow, got error: %s", err))
+		return
+	}
+}
+
+func (r *IncidentWorkflowResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
+}
+
+func (r *IncidentWorkflowResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	client, ok := req.ProviderData.(*client.ClientWithResponses)
+	if !ok {
+		resp.Diagnostics.AddError(
+			"Unexpected Resource Configure Type",
+			fmt.Sprintf("Expected *client.Client, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+		)
+
+		return
+	}
+
+	r.client = client
+}
+
+func (r *IncidentWorkflowResource) buildModel(workflow client.Workflow) *IncidentWorkflowResourceModel {
+	model := &IncidentWorkflowResourceModel{
+		ID:      types.StringValue(workflow.Id),
+		Name:    types.StringValue(workflow.Name),
+		Trigger: types.StringValue(workflow.Trigger.Name),
+	}
+	if workflow.Folder != nil {
+		model.Folder = types.StringValue(*workflow.Folder)
+	}
+	if workflow.TerraformRepoUrl != nil {
+		model.TerraformRepoURL = types.StringValue(*workflow.TerraformRepoUrl)
+	}
+	return model
+}