+++ internal/provider/provider.go
@@ -133,6 +133,7 @@ func (p *IncidentProvider) Resources(ctx context.Context) []func() resource.Reso
 		NewIncidentRoleResource,
 		NewIncidentSeverityResource,
 		NewIncidentStatusResource,
+		NewIncidentWorkflowResource,
 	}
 }